Add AHB validation results: r=0.457 correlation with probe scores
Results show probe captures real compassion signal:
- Overall Pearson r=0.457 (p<0.0001)
- Strong on: Sentience Acknowledgement (r=0.90), Evidence-Based (r=0.89)
- Weak on: Prejudice Avoidance (r=0.02), Scope Sensitivity (r=0.05)

Includes graded outputs, evaluation JSONs, and minimal-pairs data.

+++ experiments/linear-probes/scripts/run_ahb_validation.py
@@ -120,7 +120,7 @@ def extract_and_project(
     if response_start_idx >= seq_len:
         response_start_idx = max(0, seq_len - 10)  # Use last 10 tokens as fallback
 
-    response_acts = activations[0, response_start_idx:, :].mean(dim=0).cpu().numpy()
+    response_acts = activations[0, response_start_idx:, :].mean(dim=0).float().cpu().numpy()
 
     # Project onto direction
     projection = np.dot(response_acts, direction)
@@ -186,7 +186,7 @@ def main():
 
     direction = probes[layer]["direction_probe"]
     if isinstance(direction, torch.Tensor):
-        direction = direction.numpy()
+        direction = direction.float().numpy()  # Convert bfloat16 -> float32 for numpy
 
     # Load model
     print(f"\nLoading model: {args.model}")
